requirements.txt
astropy 5.0.1
collection 0.1.6
imbalanced-learn 0.9.0
imblearn 0.0
keras  2.8.0
matplotlib 3.5.1
nolds  0.5.2
numpy  1.21.4
pandas 1.4.0
py-ecg-detectors 1.2.0
scikit-learn 1.0.2
scipy 1.7.3
seaborn 0.11.2
sklearn 0.0
tensorflow 2.8.0
wfdb 2.2.1
joblib